+++ runtime/module.go
@@ -14,8 +14,7 @@ import (
 	reflectionv1 "cosmossdk.io/api/cosmos/reflection/v1"
 	"cosmossdk.io/core/appmodule"
 	"cosmossdk.io/core/comet"
-	"cosmossdk.io/core/legacy"
-	"cosmossdk.io/core/server"
+	"cosmossdk.io/core/registry"
 	"cosmossdk.io/core/store"
 	"cosmossdk.io/depinject"
 	"cosmossdk.io/depinject/appconfig"
@@ -103,7 +102,6 @@ func init() {
 			ProvideEnvironment,
 			ProvideTransientStoreService,
 			ProvideModuleManager,
-			ProvideAppVersionModifier,
 			ProvideCometService,
 		),
 		appconfig.Invoke(SetupAppBuilder),
@@ -112,7 +110,7 @@ func init() {
 
 func ProvideApp(
 	interfaceRegistry codectypes.InterfaceRegistry,
-	amino legacy.Amino,
+	amino registry.AminoRegistrar,
 	protoCodec *codec.ProtoCodec,
 ) (
 	*AppBuilder,
@@ -159,7 +157,7 @@ type AppInputs struct {
 	ModuleManager     *module.Manager
 	BaseAppOptions    []BaseAppOption
 	InterfaceRegistry codectypes.InterfaceRegistry
-	LegacyAmino       legacy.Amino
+	LegacyAmino       registry.AminoRegistrar
 	AppOptions        servertypes.AppOptions `optional:"true"` // can be nil in client wiring
 }
 
@@ -291,10 +289,6 @@ func ProvideTransientStoreService(
 	}
 
 	return transientStoreService{key: storeKey}
-}
-
-func ProvideAppVersionModifier(app *AppBuilder) server.VersionModifier {
-	return app.app
 }
 
 func ProvideCometService() comet.Service {